requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
bcrypt==4.2.1
cryptography==43.0.1
Flask-Compress==1.17
pywebpush==2.0.0
pytest==8.3.4
pytest-flask==1.3.0
